fix: park the /pricing links until the page exists (#1730)
#2284 shipped the Pricing entry in PublicHeader, the footer's Product column
and the free-core band's CTA, but `/pricing` itself ships in #1730 and does
not exist yet. The App Router prefetches every in-viewport <Link>, so the
prefetch of a route with no page returned 404 straight into the browser
console — and `e2e/smoke.spec.ts:30` asserts the landing page produces no
console diagnostics at all.

The result was main's entire merge gate going red: every Playwright smoke run
since 012d03d2 has failed on that one test, on `main` itself and therefore on
every open PR rebased onto it (#2265, the prod-deploy unblock, among them).
The failure looks unrelated to whatever PR reports it, which is what makes it
expensive.

Parked, not reverted: `publicNav.pricing`, `landing.freeCoreBadge`,
`freeCoreBody` and `freeCoreCta` stay in all three dictionaries, so #1730
restores three lines and its own nav assertions. The band keeps its badge,
headline and body — only the click is missing. The CTA cannot point somewhere
else in the meantime: `freeCoreCta` promises the price in EN/TR/DE, and a
button reading "See pricing" that lands on /features is worse than no button.

e2e/pricing-nav.spec.ts becomes e2e/free-core-band.spec.ts: the three link
tests and the phone-menu pricing test go (they assert the parked links), the
band's content assertion and its 375px overflow check stay.
