+++ src/components/landing/PublicFooter.tsx
@@ -28,7 +28,8 @@ export async function PublicFooter() {
         { href: '/for-companies', label: n.forCompanies },
         // #1732 — the price is a Product answer, not a legal one, and the
         // footer is where a visitor who scrolled past the band looks for it.
-        { href: '/pricing', label: n.pricing },
+        // Parked with the header entry until `/pricing` exists; see the note in
+        // PublicHeader.
         { href: '/projects', label: n.showcase },
         { href: '/release-notes', label: n.whatsNew },
         // The demo links to itself from its own footer — hide it there.
