+++ src/components/landing/PublicHeader.tsx
@@ -63,13 +63,16 @@ export function PublicHeader({
 
   // One array drives both the `lg` nav bar and the mobile disclosure panel
   // below, so a link added here is reachable on a phone by construction.
-  // Pricing sits next to "For companies" (#1732): the chrome offered Features /
-  // For companies / Showcase and never mentioned money, so a visitor who came
-  // for the number had nowhere to click.
+  //
+  // The Pricing entry (#1732) is parked until `/pricing` exists: #2284 shipped
+  // the link before the page, and App Router prefetches every in-viewport
+  // <Link>, so the 404 landed in the console and reded `smoke.spec.ts` — the
+  // whole merge gate — on main. `publicNav.pricing` stays in all three
+  // dictionaries, so restoring the entry is a one-line change once the page
+  // lands.
   const links = [
     { href: '/features', label: n.features },
     { href: '/for-companies', label: n.forCompanies },
-    { href: '/pricing', label: n.pricing },
     { href: '/projects', label: n.showcase },
   ];
